+++ geotrek/tourism/parsers.py
@@ -464,7 +464,7 @@ class EspritParcParser(AttachmentParserMixin, Parser):
         'name': 'nomCommercial',
         'description': 'descriptionDetaillee',
         'practical_info': 'informationsPratiques',
-        'category': 'type.0.label',
+        'category': 'type.label',
         'contact': (
             'contact.adresse',
             'contact.codePostal',
@@ -499,8 +499,8 @@ class EspritParcParser(AttachmentParserMixin, Parser):
     }
 
     m2m_fields = {
-        'type1': 'sousType.*.label',
-        'type2': 'classement.*.labelType',
+        'type1': 'sousType.label',
+        'type2': 'classement',
     }
 
     non_fields = {
@@ -557,23 +557,23 @@ class EspritParcParser(AttachmentParserMixin, Parser):
 
     def filter_type1(self, src, val):
         dst = []
-        for subval in val or []:
+        if val:
             try:
-                dst.append(TouristicContentType1.objects.get(category=self.obj.category, label=subval))
+                dst.append(TouristicContentType1.objects.get(category=self.obj.category, label=val))
             except TouristicContentType1.DoesNotExist:
                 self.add_warning(
                     _("Type 1 '{subval}' does not exist for category '{cat}'. Please add it").format(
-                        subval=subval, cat=self.obj.category.label))
+                        subval=val, cat=self.obj.category.label))
         return dst
 
     def filter_type2(self, src, val):
         dst = []
-        for subval in val or []:
+        if val:
             try:
-                dst.append(TouristicContentType2.objects.get(category=self.obj.category, label=subval))
+                dst.append(TouristicContentType2.objects.get(category=self.obj.category, label=val))
             except TouristicContentType2.DoesNotExist:
                 self.add_warning(_("Type 2 '{subval}' does not exist for category '{cat}'. Please add it").format(
-                    subval=subval, cat=self.obj.category.label))
+                    subval=val, cat=self.obj.category.label))
         return dst